requirements.txt
PyYAML
statsmodels
matplotlib
seaborn
pandas
scikit-learn
xgboost 
tensorflow == 2.17.1
ipykernel